breast-cancer-detection/requirements.txt
streamlit>=1.32.0
flask>=3.0.0
werkzeug>=3.0.0
numpy>=1.24.0
Pillow>=10.0.0
tensorflow-cpu>=2.15.0
scikit-learn>=1.3.0
imbalanced-learn>=0.11.0
matplotlib>=3.7.0
pandas>=2.0.0
gunicorn>=21.0.0
